+++ Billiard3D/Billiard3D/quat.h
@@ -6,13 +6,13 @@
 
 class quat{
 public:
-	double l;
+	float l;
 	vec v;
 
 	quat();
-	quat(double l, vec v);
+	quat(float l, vec v);
 
-    double mod();
+    float mod();
     quat normalized();
 
 	quat operator+ (quat b);
@@ -21,11 +21,11 @@ public:
 	quat operator- (quat b);
 	quat operator-= (quat b);
 
-	friend quat operator* (double a, quat b);
-	friend quat operator* (quat b, double a);
+	friend quat operator* (float a, quat b);
+	friend quat operator* (quat b, float a);
 
-	friend quat operator/ (double a, quat b);
-	friend quat operator/ (quat b, double a);
+	friend quat operator/ (float a, quat b);
+	friend quat operator/ (quat b, float a);
 
 	quat operator*(quat b);
 };
